refactor(input): rename MIDIBackend to RtMidiBackend, introduce platform dispatch alias
MIDIBackend was both the public name and the RtMidi implementation.
Separate these concerns: RtMidiBackend carries the implementation,
MIDIBackend.hpp becomes a thin dispatch header with a using alias.
No behaviour changes. InputSubsystem and all callers are untouched.

Prepares the platform backend split: PipeWireMIDIBackend (Linux),
CoreMIDIBackend (macOS), WinMMMIDIBackend (Windows).
